move project categorization to settings.gradle
Before this change we applied the cross-configuration logic for all
projects, regardless if they were used later or not. This hampered the
usefulness of configure-on-demand.
After this change, the build logic for each project is only applied if
that project is actually needed for the particular build. So if you only
call 'runIdea' in org.eclipse.xtend.idea, then the configuration for the
jenkins, web and test projects is never evaluated.